feat: implement methods for retrieving 1. available categories, languages, and regions 2. corresponding code by language or name
